api/requirements-deploy.txt
fastapi==0.115.12
uvicorn==0.34.2
python-multipart==0.0.20
python-dotenv==1.0.0
pillow==10.0.0
torch==2.7.0
torchvision==0.22.0
numpy==1.24.3
opencv-python==4.8.1.78
matplotlib==3.7.2
requests==2.32.3
